TODO: add built-in settings for wifi and cell
The original cell udev rule for triggering cell connectivity using qmicli is
removed. This is due to the fact that qmicli and mmcli have conflicts.

We should be able to configure the cell and wifi network with networkmanager and
modemmanager for now but we need to have built-in configurations for our
modem and known wifi hot spots.